+++ src/config/jwt.js
@@ -1,0 +1,4 @@
+const privateKey = 'abC123!';
+
+// node & browser
+module.exports = privateKey;