+++ src/interfaces/custom.window.ts
@@ -1,0 +1,15 @@
+export interface CustomWindow extends Window {
+    io?: any;
+}
+/*
+declare global {
+    interface Window {
+        io?: any;
+    }
+}
+
+or:
+
+declare const window: CustomWindow;
+*/
+