+++ src/utils/data-service/auth-header.js
@@ -14,7 +14,7 @@ function authHeader() {
     }
 }
 
-module.exports = authHeader;
+export default authHeader;
 
 /*
 Example: