+++ src/utils/hooks/useAuth.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import AuthContext from '@/contexts/JWTAuthContext';
+
+const useAuth = () => useContext(AuthContext);
+
+export default useAuth;