+++ src/utils/libs/auth-header.d.ts
@@ -1,6 +1,0 @@
-export = authHeader;
-/**
- * Add header authorization data
- * @returns {Object}
- */
-declare function authHeader(): any;