+++ src/utils/libs/is-admin.d.ts
@@ -1,6 +1,0 @@
-export = isAdmin;
-/**
- * Determine whether you have administrator authorization
- * @returns {Boolean}
- */
-declare function isAdmin(): boolean;