+++ src/utils/libs/jwt.d.ts
@@ -1,0 +1,8 @@
+/**
+ * JWT
+ */
+export const JWT_SECRET: "jwt-secret-key";
+export const JWT_EXPIRES_IN: number;
+export function sign(payload: any, privateKey: any, header: any): string;
+export function decode(token: any): any;
+export function verify(token: any, privateKey: any): any;